calculator/member: remove perms based on channel
Remove permissions based on the type of channel when using the
`MemberCalculator::in_channel` method. For every channel type, remove
the guild-level permissions, such as "Ban Members" and "Kick Members".
When the channel is a text channel, remove voice permissions such as
"Connect" or "Speak", and when the channel is a voice channel remove
text permissions such as "Mention Everyone" and "Send Messages".

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>