fix: skip channel overwrites if member is admin
If the member has the "Administrator" permission from the calculated
root-level permissions, then skip any calculations in the `in_channel`
method and just return all permissions.

This fixes the scenario where a member with a guild-level
"Administrator" permission has an overwrite that denies it, and then no
longer having that permission. Administrators should always have all
permissions.

A test has been added for this scenario.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>